refactor using optional functional to handle err

+++ src/main/java/ir/mvbdx/mywallet/service/impl/CategoryServiceImpl.java
@@ -9,11 +9,11 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class CategoryServiceImpl implements CategoryService {
+    public static final String CLASS_NAME = Category.class.getSimpleName();
     private final CategoryRepository categoryRepository;
 
     @Override
@@ -23,7 +23,7 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public Category findById(Long id) {
-        return categoryRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(id, Category.class.getSimpleName()));
+        return categoryRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(id, CLASS_NAME));
     }
 
     @Override
@@ -33,17 +33,15 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public Category update(Long id, Category category) {
-        categoryRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(id, Category.class.getSimpleName()));
+        categoryRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(id, CLASS_NAME));
         return categoryRepository.save(category);
     }
 
     @Override
     public void delete(Long id) {
-        Optional<Category> category = categoryRepository.findById(id);
-        if (category.isEmpty())
-            throw new EntityNotFoundException(id, Category.class.getSimpleName());
-        if (!category.get().getTransactions().isEmpty() || !category.get().getSubordinates().isEmpty())
-            throw new EntityHaveRelationException(Category.class.getSimpleName() + " " + category.get().getName());
+        Category category = categoryRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(id, CLASS_NAME));
+        if (!category.getTransactions().isEmpty() || !category.getSubordinates().isEmpty())
+            throw new EntityHaveRelationException(CLASS_NAME + " " + category.getName(), id);
         categoryRepository.deleteById(id);
     }
 }
